requirements.txt
# requirements.txt

# Web Frameworks
streamlit>=1.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Azure OpenAI
openai>=1.12.0
python-dotenv>=1.0.0

# Vector Search & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.3.1

# Utilities
pydantic>=2.5.3
httpx>=0.26.0
pyyaml>=6.0.1
pandas>=2.2.0
numpy>=1.26.3

# Background Tasks
asyncio>=3.4.3